Modified MovieParameters builder
